Added a SparseNDistanceFunction - much faster
We added a SparseNDistanceFunction implementation which holds a map
from all values which the `f` maps to a non-zero output to that output.

It's pretty naive, but ends up being much faster (~10x in the 2D case)

We also suspect that it will scale better to higher dimensions.
